requirements.txt
seleniumbase==4.*
webdriver-manager==4.*
python-dotenv==1.*
pyotp==2.*
selenium>=4.20.0
undetected-chromedriver>=3.5.6